+++ tiledb/sm/rest/rest_client_remote.cc
@@ -1516,7 +1516,7 @@ RestClientRemote::post_consolidation_plan_from_rest(
       serialization_type_, returned_data);
 }
 
-const RestCapabilities& RestClientRemote::get_capabilities_from_rest() {
+const RestCapabilities& RestClientRemote::get_capabilities_from_rest() const {
   // Return early if already detected REST capabilities for this REST client.
   if (rest_capabilities_.detected_) {
     return rest_capabilities_;
@@ -1524,8 +1524,9 @@ const RestCapabilities& RestClientRemote::get_capabilities_from_rest() {
 
   // Init curl and form the URL
   Curl curlc(logger_);
+  std::unordered_map<std::string, std::string> redirect_meta;
   throw_if_not_ok(curlc.init(
-      config_, extra_headers_, &redirect_meta_, &redirect_mtx_, false));
+      config_, extra_headers_, &redirect_meta, &redirect_mtx_, false));
   const std::string url = rest_server_ + "/v4/capabilities";
 
   Buffer data;
