+++ tiledb/sm/serialization/rest_capabilities.cc
@@ -1,5 +1,5 @@
 /**
- * @file rest_version.cc
+ * @file rest_capabilities.cc
  *
  * @section LICENSE
  *
@@ -43,7 +43,7 @@
 #include "tiledb/sm/enums/serialization_type.h"
 #include "tiledb/sm/rest/rest_client.h"
 #include "tiledb/sm/rest/rest_client_remote.h"
-#include "tiledb/sm/serialization/rest_version.h"
+#include "tiledb/sm/serialization/rest_capabilities.h"
 #include "tiledb/sm/storage_manager/context.h"
 
 using namespace tiledb::common;
